interrupts: use loader IRQ handler for Jailhouse
There have been issues early on when installing our own interrupt
handler. The main thread seemed to simply stop executing after the
interrupt handler has run for the first time. Interrupts still worked,
and the handler was called and ran successfully every time, but the
return address was always the same, as if the main thread had simply
stopped executing.

Not sure what the issue was with that, but we are now using a workaround
that bounces off the handler in the loader program, which always worked
for some reason.

There have been numerous changes to the loader, the Jailhouse cell
config etc. since, so this may be worth revisiting, but for now I'll
leave it the way it works.