mmu: limit RAM size detection to 2GB minus 256M
If we want to pass pointers to the Linux cell, we need to make sure that
the Linux cell can map our used address space, in its entirety, to the
same virtual address in its address space. Unfortunately, there is not
enough address space available in Linux to do that on devices with 2 GBs
of RAM, if only by a small margin.
Since nobody is likely to actually require all 2 GBs of RAM for BASIC,
we just leave the last 256 MB unused and go on with our day.